+++ apps/submitter/src/validation/schemas/happyTxState.ts
@@ -1,0 +1,45 @@
+import { z } from "zod"
+import { EntryPointStatus, SubmitterErrorStatus } from "#src/tmp/interface/status"
+import { happyReceiptSchema } from "./happyReceipt"
+import { simulationResultSchema } from "./simulationResult"
+
+const HappyTxStateSubmitterErrorSchema = z.object({
+    status: z.enum([
+        SubmitterErrorStatus.BufferExceeded,
+        SubmitterErrorStatus.OverCapacity,
+        SubmitterErrorStatus.UnexpectedError,
+        SubmitterErrorStatus.SimulationTimeout,
+    ]),
+    included: z.literal(undefined),
+})
+
+const HappyTxStateEntryPointErrorSchema = z.object({
+    status: z.enum([
+        // EntryPointStatus
+        EntryPointStatus.Success,
+        EntryPointStatus.ValidationReverted,
+        EntryPointStatus.ValidationFailed,
+        EntryPointStatus.ExecuteReverted,
+        EntryPointStatus.ExecuteFailed,
+        EntryPointStatus.PaymentReverted,
+        EntryPointStatus.PaymentFailed,
+        EntryPointStatus.UnexpectedReverted,
+        // SubmitterErrorSimulationMaybeAvailable
+        SubmitterErrorStatus.SubmitTimeout,
+        SubmitterErrorStatus.ReceiptTimeout,
+    ]),
+    included: z.literal(false),
+    simulation: simulationResultSchema.optional(),
+})
+
+const HappyTxStateSuccessSchema = z.object({
+    status: z.literal(EntryPointStatus.Success).openapi({ example: EntryPointStatus.Success }),
+    included: z.literal(true).openapi({ example: true }),
+    receipt: happyReceiptSchema,
+})
+
+export const happyTxStateSchema = z.discriminatedUnion("included", [
+    HappyTxStateSuccessSchema,
+    HappyTxStateEntryPointErrorSchema,
+    HappyTxStateSubmitterErrorSchema,
+])
